Renamed reducers-actions to state
