+++ test/components/DepositWidget.components.test.tsx
@@ -6,7 +6,7 @@ import { Row, RowProps } from 'components/DepositWidget/Row'
 
 import { ZERO, ONE, TEN } from 'const'
 import { Network, TokenBalanceDetails } from 'types'
-import { TokenLocalState } from 'reducers-actions'
+import { TokenLocalState } from 'state'
 import { createFlux } from '../data'
 
 const fakeRowState: Record<keyof TokenLocalState, boolean> = {
